Move MouseZoomOrScrollListener to own file
